+++ src/inspect_ai/_view/www/src/App.mjs
@@ -37,7 +37,10 @@ import {
 } from "./samples/SamplesDescriptor.mjs";
 import { byEpoch, bySample, sortSamples } from "./samples/tools/SortFilter.mjs";
 import { resolveAttachments } from "./utils/attachments.mjs";
-import { filterFnForType } from "./samples/tools/filters.mjs";
+import {
+  addFragmentToFilter,
+  filterExpression,
+} from "./samples/tools/filters.mjs";
 
 import {
   kEvalWorkspaceTabId,
@@ -162,6 +165,7 @@ export function App({
   const [score, setScore] = useState(initialState?.score);
 
   // Re-filter the samples
+  const [filterError, setFilterError] = useState(initialState?.filterError);
   const [filteredSamples, setFilteredSamples] = useState(
     initialState?.filteredSamples || [],
   );
@@ -169,6 +173,14 @@ export function App({
   const [groupByOrder, setGroupByOrder] = useState(
     initialState?.groupByOrder || "asc",
   );
+
+  const addToFilterExpression = (fragment) => {
+    setFilter(addFragmentToFilter(filter, fragment));
+    const filterInput = document.getElementById("sample-filter-input");
+    if (filterInput) {
+      filterInput.focus();
+    }
+  };
 
   const afterBodyElements = [];
   const saveState = useCallback(() => {
@@ -306,6 +318,7 @@ export function App({
 
   useEffect(() => {
     const samples = selectedLog?.contents?.sampleSummaries || [];
+    var newFilterError = undefined;
     const filtered = samples.filter((sample) => {
       // Filter by epoch if specified
       if (epoch && epoch !== "all") {
@@ -315,9 +328,14 @@ export function App({
       }
 
       // Apply the filter
-      const filterFn = filterFnForType(filter);
-      if (filterFn && filter.value) {
-        return filterFn(samplesDescriptor, sample, filter.value);
+      if (filter.value) {
+        const { matches, error } = filterExpression(
+          evalDescriptor,
+          sample,
+          filter.value,
+        );
+        newFilterError ||= error;
+        return matches;
       } else {
         return true;
       }
@@ -336,6 +354,7 @@ export function App({
       }
     }
 
+    setFilterError(newFilterError);
     setFilteredSamples(sorted);
     setGroupBy(grouping);
     setGroupByOrder(order);
@@ -913,7 +932,9 @@ export function App({
               epoch=${epoch}
               setEpoch=${setEpoch}
               filter=${filter}
+              filterError=${filterError}
               setFilter=${setFilter}
+              addToFilterExpression=${addToFilterExpression}
               score=${score}
               setScore=${setScore}
               scores=${scores}