+++ src/inspect_ai/_view/www/src/Types.mjs
@@ -26,7 +26,6 @@
 /**
  * @typedef {Object} ScoreFilter
  * @property {string} [value]
- * @property {string} [type]
  */
 
 /**